+++ lib/ReactViewModels/TerriaRouting.ts
@@ -1,8 +1,0 @@
-// Route paths for the app.
-
-export const ROOT_ROUTE = "/";
-
-export const CATALOG_ROUTE = `${ROOT_ROUTE}catalog/`;
-// var CATALOG_ROUTE = "/catalog/";
-
-export const CATALOG_MEMBER_ROUTE = `${CATALOG_ROUTE}:catalogMemberId`;