+++ lib/ReactViews/Preview/DataPreview.jsx
@@ -1,31 +1,19 @@
 "use strict";
-import React from "react";
-import { withRouter } from "react-router-dom";
-import { Helmet } from "react-helmet";
 
 // import Chart from "../Custom/Chart/Chart";
-import Description, { getMetaDescriptionSummary } from "./Description";
+import Description from "./Description";
 import GroupPreview from "./GroupPreview";
 import InvokeFunction from "../Analytics/InvokeFunction";
 import MappablePreview from "./MappablePreview";
+import React from "react";
 import createReactClass from "create-react-class";
 import PropTypes from "prop-types";
 import Styles from "./data-preview.scss";
-
-import { withTranslation } from "react-i18next";
+import { withTranslation, Trans } from "react-i18next";
 import { observer } from "mobx-react";
 import { runInAction } from "mobx";
 import ReferenceMixin from "../../ModelMixins/ReferenceMixin";
 import Loader from "../Loader";
-import Box from "../../Styled/Box";
-import Button from "../../Styled/Button";
-import { ROOT_ROUTE } from "../../ReactViewModels/TerriaRouting";
-
-const prerenderEnd = () => {
-  if (document && document.dispatchEvent) {
-    document.dispatchEvent(new Event("prerender-end"));
-  }
-};
 
 /**
  * Data preview section, for the preview map see DataPreviewMap
@@ -38,47 +26,7 @@ const DataPreview = observer(
       terria: PropTypes.object.isRequired,
       viewState: PropTypes.object,
       previewed: PropTypes.object,
-      location: PropTypes.object.isRequired,
       t: PropTypes.func.isRequired
-    },
-
-    getInitialState() {
-      return {
-        // For prerendering & expanding parent catalog on first load
-        loaded: false
-      };
-    },
-
-    componentDidMount() {
-      // Make sure our prerenderer doesn't stall waiting forever
-      setTimeout(prerenderEnd, 10000);
-      this.checkEnableWithParents();
-    },
-
-    componentDidUpdate() {
-      this.checkEnableWithParents();
-    },
-
-    checkEnableWithParents() {
-      if (
-        !this.state.loaded &&
-        this.props.previewed &&
-        !this.props.previewed.isLoading
-      ) {
-        // also enable parents for the first load ever,
-        // to ensure that the catalog hierarchy is visible as users can land
-        // directly on a previewed item now via URL-routing only, without share link
-        this.props.previewed.parent &&
-          this.props.previewed.parent.enableWithParents &&
-          this.props.previewed.parent.enableWithParents();
-
-        prerenderEnd();
-
-        this.setState({ loaded: true });
-      } else if (!this.state.loaded && !this.props.previewed) {
-        // we're on /catalog/ without a previewed item, immediately prerender end
-        prerenderEnd();
-      }
     },
 
     backToMap() {
@@ -97,33 +45,14 @@ const DataPreview = observer(
         }
         previewed = previewed.target;
       }
-      const appBaseUrl = this.props.terria.configParameters.appBaseUrl;
-      const pathname = this.props.location && this.props.location.pathname;
-      const pathForCanonical =
-        pathname && pathname.charAt(0) === "/" ? pathname.slice(1) : pathname;
+
       let chartData;
       if (previewed && !previewed.isMappable && previewed.tableStructure) {
         chartData = previewed.chartData();
       }
+
       return (
         <div className={Styles.preview}>
-          <If condition={previewed}>
-            <Helmet>
-              <title>
-                {previewed.name} - {this.props.terria.appName}
-              </title>
-              <meta
-                name="description"
-                content={getMetaDescriptionSummary(previewed)}
-              />
-              <If condition={appBaseUrl && pathForCanonical}>
-                <link
-                  rel="canonical"
-                  href={`${appBaseUrl}${pathForCanonical}`}
-                />
-              </If>
-            </Helmet>
-          </If>
           <Choose>
             <When condition={previewed && previewed.isMappable}>
               <div className={Styles.previewInner}>
@@ -170,13 +99,16 @@ const DataPreview = observer(
             </When>
             <Otherwise>
               <div className={Styles.placeholder}>
-                <p>{t("preview.selectToPreview")}</p>
-                <p>{t("preview.or")}</p>
-                <Box centered>
-                  <Button secondary shortMinHeight renderAsLink to={ROOT_ROUTE}>
-                    {t("preview.goToTheMap")}
-                  </Button>
-                </Box>
+                <Trans i18nKey="preview.selectToPreview">
+                  <p>Select a dataset to see a preview</p>
+                  <p>- OR -</p>
+                  <button
+                    className={Styles.btnBackToMap}
+                    onClick={this.backToMap}
+                  >
+                    Go to the map
+                  </button>
+                </Trans>
               </div>
             </Otherwise>
           </Choose>
@@ -207,4 +139,4 @@ const DataPreview = observer(
   })
 );
 
-module.exports = withRouter(withTranslation()(DataPreview));
+module.exports = withTranslation()(DataPreview);