+++ lib/ReactViewModels/ViewState.ts
@@ -7,8 +7,6 @@ import {
   runInAction
 } from "mobx";
 import { Ref } from "react";
-import { History } from "history";
-
 import clone from "terriajs-cesium/Source/Core/clone";
 import defined from "terriajs-cesium/Source/Core/defined";
 import CesiumEvent from "terriajs-cesium/Source/Core/Event";
@@ -70,7 +68,6 @@ export default class ViewState {
   readonly searchState: SearchState;
   readonly terria: Terria;
   readonly relativePosition = RelativePosition;
-  history: History | undefined;
 
   @observable previewedItem: BaseModel | undefined;
   @observable userDataPreviewedItem: BaseModel | undefined;