+++ lib/ReactViews/Preview/DataPreviewSections.jsx
@@ -64,7 +64,7 @@ const DataPreviewSections = observer(
             <Choose>
               <When condition={item.content !== undefined}>
                 <div key={i}>
-                  <h2 className={Styles.subHeading}>{item.name}</h2>
+                  <h4 className={Styles.h4}>{item.name}</h4>
                   {parseCustomMarkdownToReact(
                     Mustache.render(item.content, metadataItem),
                     {
@@ -74,7 +74,7 @@ const DataPreviewSections = observer(
                 </div>
               </When>
               <When condition={item.contentAsObject !== undefined}>
-                <h2 className={Styles.subHeading}>{item.name}</h2>
+                <h4 className={Styles.h4}>{item.name}</h4>
                 <MetadataTable metadataItem={item.contentAsObject} />
               </When>
             </Choose>