+++ lib/ReactViews/Preview/data-preview.scss.d.ts
@@ -12,7 +12,6 @@ interface CssExports {
   'previewChart': string;
   'previewInner': string;
   'preview__inner': string;
-  'subHeading': string;
 }
 declare var cssExports: CssExports;
 export = cssExports;