+++ CHANGES.md
@@ -5,6 +5,8 @@ Change Log
 
 
 #### next release (8.0.0-alpha.60)
+* Fix WMS legend for default styles.
+* Request transparent legend from GeoServer.
 * [The next improvement]
 
 #### 8.0.0-alpha.59
@@ -21,7 +23,6 @@ Change Log
   allowing TerriaMap to hit certain routes)
 * Update `WorkbenchButton` to allow for links rather than buttons, including
   changing About Data to a link
-* [The next improvement]
 
 #### 8.0.0-alpha.58
 * Add `FeatureInfoTraits` to `ArcGisMapServerCatalogItem`