+++ CHANGES.md
@@ -3,11 +3,16 @@ Change Log
 
 ### MobX Development
 
+#### next release
 
-#### next release (8.0.0-alpha.60)
+#### 8.0.0-alpha.60
 * Fix WMS legend for default styles.
 * Request transparent legend from GeoServer.
-* [The next improvement]
+* Reverted the following due to various issues with datasets:
+    * Add basic routing support
+    * Add better page titles when on various routes of the application
+    * Add prerendering support on `/catalog/` routes (via `prerender-end` event & 
+      allowing TerriaMap to hit certain routes)
 
 #### 8.0.0-alpha.59
 * Update magda error message